pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jira-issue-tracker-server"
version = "0.2.0"
description = "A Kivy-based desktop application for tracking Jira issues with custom JQL queries"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Josh Simnitt", email = "josh.simnitt@example.com"}
]
keywords = ["jira", "issue-tracker", "desktop-app", "kivy", "jql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Bug Tracking",
]
dependencies = [
    "kivy>=2.2.0",
    "kivymd>=1.1.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "urllib3>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/star7js/jira-issue-tracker-server"
Repository = "https://github.com/star7js/jira-issue-tracker-server"
Issues = "https://github.com/star7js/jira-issue-tracker-server/issues"

[project.scripts]
jira-tracker = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*.py"]
exclude = ["tests*", "test_*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["test_*.py"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
